renames web_client/Components to web_client/components
