+++ webApp/web_client/components/BroadcastEntry.js
