+++ webApp/web_client/components/BroadcastLiveView.js
