+++ webApp/web_client/components/BroadcastLiveViewAUX.js
