+++ webApp/web_client/components/BroadcastLiveViewSC.js
