+++ webApp/web_client/components/BroadcastSCEntry.js
