+++ webApp/web_client/components/NavBar.js
