+++ webApp/web_client/components/NavBarView.js
