+++ webApp/web_client/components/Stream.js
