+++ webApp/web_client/components/StreamEntry.js
