+++ webApp/web_client/components/StreamLiveView.js
