+++ webApp/web_client/components/StreamLiveViewLinkbar.js
